refact:  add permssion check 1. go-jsonrpc support field reflect 2. go-ipfs-cmds support auth check in hanlder 3. venus client support  dynamically specifies whether to expose to external calls 4. venus client code can generate by tools

+++ cmd/miner_actor.go
@@ -49,7 +49,7 @@ var actorSetAddrsCmd = &cmds.Command{
 		cmds.StringArg("address", true, false, "Address of miner to show"),
 	},
 	Options: []cmds.Option{
-		cmds.Int64Option("gas-limit", "set gas limit").WithDefault(0),
+		cmds.Int64Option("gas-limit", "set gas limit").WithDefault(int64(0)),
 		cmds.StringsOption("addrs", "set addresses"),
 	},
 	Run: func(req *cmds.Request, re cmds.ResponseEmitter, env cmds.Environment) error {
